requirements.txt
tensorflow
numpy
opencv-python
pillow
scipy
datasets
